manque plus que hexa et octal

+++ ft_short.c
@@ -6,7 +6,7 @@
 /*   By: pgrassin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/06/07 16:18:23 by pgrassin          #+#    #+#             */
-/*   Updated: 2016/06/07 16:23:46 by pgrassin         ###   ########.fr       */
+/*   Updated: 2016/06/09 15:46:40 by pgrassin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,11 +16,11 @@
 
 int		ft_short(t_module *m, va_list args)
 {
-	intmax_t	val;
+	short	val;
 	int			i;
 
-	val = (intmax_t)va_arg(args, unsigned long);
-	i = ft_int_init(m, val, 10, "0123456789");
+	val = (short)va_arg(args, int);
+	i = ft_int_init(m, (intmax_t)val, 10, "0123456789");
 	if (m->flag.moins)
 		return (ft_int_moins(m, val, i, m->prec));
 	else
@@ -29,11 +29,11 @@ int		ft_short(t_module *m, va_list args)
 
 int		ft_ushort(t_module *m, va_list args)
 {
-	intmax_t	val;
-	int			i;
+	unsigned short	val;
+	int				i;
 
-	val = (intmax_t)va_arg(args, unsigned long);
-	i = ft_int_init(m, val, 10, "0123456789");
+	val = (unsigned short)va_arg(args,unsigned int);
+	i = ft_int_init(m, (intmax_t)val, 10, "0123456789");
 	if (m->flag.moins)
 		return (ft_int_moins(m, val, i, m->prec));
 	else
